Refactor template generation out of UI
