+++ migaz/source/MIGAZ/Generator/CloudServiceVM.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MIGAZ.Generator
+{
+    public class CloudServiceVM
+    {
+        public string CloudService { get; set; }
+        public string VirtualMachine { get; set; }
+    }
+}
